Fix for #1648

+++ src/main/scala/scorex/core/transaction/state/StateFeature.scala
@@ -12,7 +12,7 @@ trait StateFeature
   * Instance of this trait supports stateful validation of any transaction
   */
 trait TransactionValidation extends StateFeature {
-  def validateWithCost(tx: ErgoTransaction, maxTxCost: Long): Try[Long]
+  def validateWithCost(tx: ErgoTransaction, maxTxCost: Int): Try[Int]
 }
 
 object TransactionValidation {
